+++ core/src/main/kotlin/siksha/wafflestudio/core/domain/common/exception/community/CommunityException.kt
@@ -1,0 +1,6 @@
+package siksha.wafflestudio.core.domain.common.exception.community
+
+import org.springframework.http.HttpStatus
+import siksha.wafflestudio.core.domain.common.exception.SikshaException
+
+sealed class CommunityException(httpStatus: HttpStatus, errorMessage: String) : SikshaException(httpStatus, errorMessage)
